# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chimera Analytics API
  description: |
    REST API for the Chimera Multi-Agent Analytics System.
    
    The Chimera API provides programmatic access to Zcash analytics capabilities including:
    - Natural language query submission
    - Report retrieval and export
    - Dashboard configuration
    - Alert management
    - Metrics data access
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: Include in `Authorization: Bearer <token>` header
    - **API Key**: Include in `X-API-Key` header
    
    ## Rate Limiting
    
    API requests are rate limited to 100 requests per minute per user/API key.
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
    
  version: 1.0.0
  contact:
    name: Chimera API Support
    email: support@chimera-analytics.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.chimera-analytics.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/dashboards:
    post:
      summary: Create dashboard
      description: Create a new dashboard configuration
      tags:
        - Dashboards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardCreate'
      responses:
        '201':
          description: Dashboard created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

    get:
      summary: List dashboards
      description: List user's dashboards
      tags:
        - Dashboards
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Dashboard list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardListResponse'
  /api/dashboards/{id}:
    get:
      summary: Get dashboard
      description: Retrieve dashboard configuration by ID
      tags:
        - Dashboards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dashboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

    put:
      summary: Update dashboard
      description: Update dashboard configuration
      tags:
        - Dashboards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardUpdate'
      responses:
        '200':
          description: Dashboard updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

    delete:
      summary: Delete dashboard
      description: Delete a dashboard
      tags:
        - Dashboards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Dashboard deleted successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
  schemas:
...
    Dashboard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        layout:
          type: object
        widgets:
          type: array
          items:
            type: object
        refreshInterval:
          type: integer
        shared:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DashboardCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        layout:
          type: object
        widgets:
          type: array
          items:
            type: object
        refreshInterval:
          type: integer
    DashboardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        layout:
          type: object
        widgets:
          type: array
        refreshInterval:
          type: integer
    DashboardListResponse:
      type: object
      properties:
        dashboards:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...